+++ src/KolibriLib/graphic.cpp
@@ -20,8 +20,8 @@ void graphic::DrawPoint(const Coord &position, const unsigned &size, const Color
 
 void graphic::DrawCircle(const Coord &coord, unsigned Radius, const Colors::Color &color)
 {
-	buf2d::buffer src({Radius * 2U, Radius * 2U}, buf2d::BPP::RGBA);
-	buf2d::buffer dst({Radius * 2U, Radius * 2U}, buf2d::BPP::RGB);
+	buf2d::buffer src({static_cast<int>(Radius * 2U), static_cast<int>(Radius * 2U)}, buf2d::BPP::RGBA);
+	buf2d::buffer dst({static_cast<int>(Radius * 2U), static_cast<int>(Radius * 2U)}, buf2d::BPP::RGB);
 
 	buf2d::ApplyTrasparency(&dst, &src);
